requirements.txt
fastapi
uvicorn
motor
pymongo
pydantic
pydantic-settings
python-dotenv
email-validator
